Refactor: move apiHook to another directory
